my queries section added all queries section functionality and ui improved
